Enhancements
Add new export image style (SSRSpeed v2.5+ result table).
Add total used time show for new export style.
Add export image color customization.
Add more log detail.
Add more preference items.
Fix Shadowsocks plugin detect in Clash and Surge parser.
Clean up unused files.
Optimize codes.

+++ misc.h
@@ -33,7 +33,7 @@ struct nodeInfo
     string pkLoss = "100.00%";
     int rawPing[6] = {};
     string avgPing = "0.00";
-    double rawSitePing[6] = {};
+    int rawSitePing[6] = {};
     string sitePing = "0.00";
     string traffic;
     geoIPInfo inboundGeoIP;
